+++ src/main/java/com/naranghiking/mtn/service/MtnService.java
@@ -8,11 +8,11 @@ public interface MtnService {
 
     List<Mtn> selectAll();
 
-    Mtn select(Integer id);
+    Mtn select(Long id);
 
     int insert(Mtn mtn);
 
     int update(Mtn mtn);
 
-    void delete(Integer id);
+    void delete(Long id);
 }
